requirements.txt
pandas==1.1.2
python-dateutil==2.8.1
coverage==5.3
pytest==6.1.1
pylint==2.6.0
